requirements.txt
appdirs==1.4.2
beautifulsoup4==4.5.3
boto==2.46.1
brotlipy==0.6.0
cffi==1.9.1
Collectfast==0.2.3
defusedxml==0.4.1
dj-database-url==0.4.2
Django==1.10.5
django-allauth==0.30.0
django-appconf==1.0.2
django-classy-tags==0.8.0
django-cleanup==0.4.2
django-crispy-forms==1.6.1
django-easy-maps==0.9.3
django-redis-cache==1.6.5
django-registration==2.2
django-smartfields==1.0.9
django-storages-redux==1.3
django-widget-tweaks==1.4.1
geopy==1.11.0
glob2==0.5
gunicorn==19.6.0
lxml==3.7.3
oauthlib==2.0.1
olefile==0.44
packaging==16.8
Pillow==4.0.0
psycopg2==2.6.2
pycparser==2.17
pyparsing==2.1.10
python-dateutil==2.6.0
python-http-client==2.2.1
python3-openid==3.0.10
pytz==2016.10
redis==2.10.5
requests==2.13.0
requests-oauthlib==0.7.0
sendgrid==3.6.3
six==1.10.0
virtualenv==15.1.0
Wand==0.4.4
whitenoise==3.3.0
